example_main shows how to use the Puzzle class from Puzzle.py
requirements.txt has been updated with the required dependencies.

Changed pyproject.toml to ignore another whitespace error (black would format, then flake8 would throw this error)

project has been updated to conform with naming and docstring conventions... maybe....

Separated puzzle classes into individual puzzle files
