+++ pyproject.toml
@@ -16,5 +16,6 @@ ignore = [
     # Docstring Content
     "D400", "D401", "D402", "D404", "D405", "D406", "D407", "D408", "D409", "D410", "D411", "D412", "D413", "D414", "D416", "D417",
     # Comments
-    "E266",
+    "E266", "E203"
+    # Whitespace
 ]
